Add TreeTableHelper base class
